analyzer/requirements.txt
python-chess==1.999
psycopg2-binary==2.9.9